Improved sql task support.

+++ src/main/groovy/com/radcortez/gradle/plugin/openjpa/sql/SqlExtension.groovy
@@ -1,8 +1,7 @@
 package com.radcortez.gradle.plugin.openjpa.sql
 
+import com.radcortez.gradle.plugin.openjpa.OpenJpaExtension
 import org.gradle.api.Project
-import org.gradle.api.tasks.Input
-import org.gradle.api.tasks.Optional
 
 import java.nio.file.Paths
 
@@ -14,24 +13,29 @@ import java.nio.file.Paths
 class SqlExtension {
     Project project
 
-    @Input
     String connectionDriverName
-    @Input
-    @Optional
     String schemaAction = "build"
-    @Input
-    @Optional
     String sqlFile = "build/database.sql"
 
     SqlExtension(final Project project) {
         this.project = project
 
         project.afterEvaluate() {
+            OpenJpaExtension configuration = project.extensions.findByType(OpenJpaExtension)
+
             project.tasks.clean.doLast {
                 cleanSqlOutputFolder()
             }
 
-            project.tasks.sql.dependsOn project.tasks.classes
+            def runSql = false
+            project.tasks.classes.doLast {
+                runSql = true
+            }
+
+            project.tasks.sql.onlyIf {
+                runSql && configuration.generateSql
+            }
+            project.tasks.classes.finalizedBy(project.tasks.sql)
         }
     }
 
@@ -42,14 +46,17 @@ class SqlExtension {
     void cleanSqlOutputFolder() {
         def sqlFile = project.file(getSqlFile()).parentFile
         def path = Paths.get(sqlFile.path)
-        sqlFile.deleteDir()
+
+        if (sqlFile.isFile()) {
+            sqlFile.deleteDir()
+        }
 
         while (path.parent != Paths.get(project.rootDir.path)) {
             path = path.parent
             if (path.toFile().exists() && path.toFile().listFiles().length == 0) {
                 path.deleteDir()
             } else {
-                break;
+                break
             }
         }
     }
